ECB - 138 Shift to new auth service
