+++ src/app/layout/header/header.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthService } from 'src/app/services/auth/auth.service';
 import { UserService } from 'src/app/services/user/user.service';
-import { MatToolbar, MatToolbarRow } from '@angular/material/toolbar';
 import 'bootstrap/dist/js/bootstrap.bundle';
+import { AuthService } from 'src/app/services/auth/auth.service';
 
 @Component({
   selector: 'app-header',
