+++ src/app/services/interceptor/jwt.token.interceptor.spec.ts
@@ -4,7 +4,7 @@ import {
   HttpTestingController,
 } from '@angular/common/http/testing';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { AuthService } from '../auth/auth.service';
+import { AuthService } from '../authOld/auth.service';
 import { JwtTokenInterceptor } from './jwt.token.interceptor';
 
 describe(`AuthHttpInterceptor`, () => {
